+++ src/empty_struct.jl
@@ -2,10 +2,10 @@
 # Returns an empty structure
 #
 
-function empty_struct(X :: DataType)
+function empty_struct(X::DataType)
   list = []
   for type in X.types
-    if type == Int64
+    if type == Int
       push!(list,0)
     end
     if type == String
