+++ src/parse_int.jl
@@ -3,12 +3,12 @@
 # or, perhaps, an hexadecimal representation string
 #
 
-function parse_int(s :: String)
+function parse_int(s::String)
   try
-    i = parse(Int64,s)
+    i = parse(Int, s)
     return i
   catch
-    i = parse(Int64,s,base=16)
+    i = parse(Int,s,base=16)
     return i
   end
   error("Could not read integer from string: \"$s\"")
